update counters types

+++ Firmware/Pump/interrupts.c
@@ -108,9 +108,9 @@ ISR(PORTC_INT0_vect, ISR_NAKED)
 /* EN_DRIVER_UC & BUT_PUSH & BUT_PULL & BUT_RESET                       */
 /************************************************************************/
 extern uint8_t but_push_counter_ms;
-extern uint8_t but_long_push_counter_ms;
+extern uint16_t but_long_push_counter_ms;
 extern uint8_t but_pull_counter_ms;
-extern uint8_t but_long_pull_counter_ms;
+extern uint16_t but_long_pull_counter_ms;
 extern uint8_t but_reset_counter_ms;
 
 ISR(PORTD_INT0_vect, ISR_NAKED)
